refactor: tooltip provider에서 tooltip 분리
