+++ app/src/@shared/ui-kit/Tooltip/index.tsx
@@ -1,1 +1,2 @@
 export * from './IconTooltip';
+export * from './Tooltip';
